% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body-links.R
\name{body_anchor_urls}
\alias{body_anchor_urls}
\title{Extract all body anchor hypertext references}
\usage{
body_anchor_urls(x, unique = FALSE, base = NULL)
}
\arguments{
\item{x}{Either a document, a node set or a single node.}

\item{unique}{if \code{TRUE}, then only unique URLs will be returned. Default: \code{FALSE}.}

\item{base}{if not \code{NULL}, all relative URLs will be converted
to absolute URLs. Default: \code{NULL}.}
}
\value{
character vector of extracted URLs; if no \code{<a>} \code{href} URLs were discovered
\code{charactrer(0)} will be returned.
}
\description{
Given a valid HTML/XML document, node set or single node, a
claracter vector of all \code{<body>} \code{<a href>} URLs will be returned.
If \code{base} is not \code{NULL} all relative URLs will be converted
to absolute URLs.
}
\examples{
x <- rvest::read_html("https://congress.gov/")
body_anchor_urls(x, unique = TRUE, base = "https://congress.gov/")
}
