% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-index-factors.R
\name{clean_index_factors}
\alias{clean_index_factors}
\title{Clean up diversity and evenness names}
\usage{
clean_index_factors(x, newline = FALSE)
}
\arguments{
\item{x}{a character vector of snake-case WSDL, Simpson and Shannon
index names (i.e. that came from a call to \code{\link[=uri_diversity]{uri_diversity()}}).}

\item{newline}{if \code{TRUE} the Title Case Versions will be separated by
a newline ("\code{\\n}") vs a space.}
}
\description{
Snake-case names are great for coding but look terrible in
tables and plots. Use this function to turn the WSDL,
Shannon and Simpson snake-case index labels into Title Case Versions
of themselves.
}
