% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uri-diversity.R
\name{uri_diversity}
\alias{uri_diversity}
\alias{url_diversity}
\title{Quantify URL diversity}
\usage{
uri_diversity(corpus, corpus_id = uuid::UUIDgenerate(),
  exception_domains = NULL)

url_diversity(corpus, corpus_id = uuid::UUIDgenerate(),
  exception_domains = NULL)
}
\arguments{
\item{corpus}{a collection (character vector) of URLs}

\item{corpus_id}{an identifier (ideally unique) for the collection; will
be generated if not provided.}

\item{exception_domains}{a character vector of domains; use this to specify domains
where the query string is important. Normally, the query string is excluded from
the canonicalized URI but in some cases (e.g. \code{youtube.com}) it is desirable
to have the query string influence the diversity computations.}
}
\value{
a data frame (tibble) with WSDL, Shannon and Simpson diversity indices for
canonical URIs and hostnames.
}
\description{
Compute WSDL Diversity Index, Shannon's evenness index, and Simpson's diversity index
for a corpus (collection) of URLs.
}
\note{
Algorithm creator: \href{https://twitter.com/acnwala}{Alexander C. Nwala}
}
\examples{
collection <- readLines(system.file("extdat", "corpus.txt", package = "urldiversity"))
uri_diversity(collection)
}
\references{
\url{http://ws-dl.blogspot.com/2018/05/2018-05-04-exploration-of-url-diversity.html}
}
\author{
Alexander Nwala (anwala@cs.odu.edu); Bob Rudis (bob@rud.is)
}
