% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urldiversity-package.R
\docType{package}
\name{urldiversity}
\alias{urldiversity}
\alias{urldiversity-package}
\title{Quantify 'URL' Diversity and Apply Popular Biodiversity Indices to a 'URL' Collection}
\description{
Methods are provided to compute the 'WSDL Diversity Index' \url{http://ws-dl.blogspot.com/2018/05/2018-05-04-exploration-of-url-diversity.html}
along with selected biodiversity indidces to a corpus (collection) of 'URLs'.
}
\references{
\url{http://ws-dl.blogspot.com/2018/05/2018-05-04-exploration-of-url-diversity.html}
}
\author{
Alexander Nwala (anwala@cs.odu.edu); Bob Rudis (bob@rud.is)
}
